Fix leading zeroes bugs (#3763)
* remove two unneeded mut's

* ensure correct size for leading_zeros()

+++ core/src/pow/cuckarood.rs
@@ -65,8 +65,8 @@ impl PoWContext for CuckaroodContext {
 		let mut ndir = vec![0usize; 2];
 		let mut xor0: u64 = 0;
 		let mut xor1: u64 = 0;
-		let mask = u64::MAX >> size.leading_zeros(); // round size up to 2-power - 1
-											 // the next two arrays form a linked list of nodes with matching bits 4..0|dir
+		let mask = u64::MAX >> (size as u64).leading_zeros(); // round size up to 2-power - 1
+													  // the next two arrays form a linked list of nodes with matching bits 4..0|dir
 		let mut headu = vec![2 * size; 1 + mask as usize];
 		let mut headv = vec![2 * size; 1 + mask as usize];
 		let mut prev = vec![0usize; 2 * size];
